test(fetch): add response cloning browser test
